+++ languages/javascript/src/shared/Transport/MockTransport.mjs
@@ -41,13 +41,13 @@ function send(message) {
     return
   }
 
-  if (json.method) {
-    let [module, method] = json.method.split('.')
+  //let [module, method] = json.method.split('.')
 
-    if (testHarness && testHarness.onSend) {
-      testHarness.onSend(module, method, json.params, json.id)
-    }
-  
+  if (testHarness && testHarness.send) {
+    testHarness.send(message)
+  }
+
+  if (json.method) {
     if (mock)
       handle(json)
     else
@@ -55,23 +55,29 @@ function send(message) {
   }
   else if (json.id !== undefined && requests[json.id]) {
     const promise = requests[json.id]
-    if (json.result !== undefined) {
-      promise.resolve(json.result)
-    }
-    else {
+    if (json.error !== undefined) {
       promise.reject(json.error)
     }
-  }
+    else {
+      promise.resolve(json.result)
+    }
 
+    delete requests[json.id]
+  }
 }
 
 function handle(json) {
   let result
   try {
     result = getResult(json.method, json.params)
+    setTimeout(() => callback(JSON.stringify({
+      jsonrpc: '2.0',
+      result: result,
+      id: json.id
+    })))
   }
   catch (error) {
-    setTimeout(() => callback(JSON.stringify({ 
+    setTimeout(() => callback(JSON.stringify({
       jsonrpc: '2.0',
       error: {
         code: -32602,
@@ -80,12 +86,6 @@ function handle(json) {
       id: json.id
     })))
   }
-
-  setTimeout(() => callback(JSON.stringify({ 
-    jsonrpc: '2.0',
-    result: result,
-    id: json.id
-  })))
 }
 
 function receive(_callback) {
@@ -114,16 +114,21 @@ function event(module, event, value) {
   }))
 }
 
+function receiveMessage (message) {
+  callback(message);
+}
+
 let id = 0
 const requests = []
 
 function request(method, params) {
+  const requestId = id++;
   const promise = new Promise( (resolve, reject) => {
-    requests[id] = { resolve, reject }
+    requests[requestId] = { resolve, reject }
   })
   callback(JSON.stringify({ 
     jsonrpc: '2.0',
-    id: id,
+    id: requestId,
     method: `${method}`,
     params: params
   }))
@@ -141,6 +146,7 @@ function dotGrab(obj = {}, key) {
 }
 
 function getResult(method, params) {
+  
   let api = dotGrab(mock, method)
 
   if (method.match(/^[a-zA-Z]+\.on[A-Za-z]+$/)) {
@@ -156,7 +162,9 @@ function getResult(method, params) {
       result = null
     }
     return result
-  } else return api
+  } 
+  else 
+    return api
 }
 
 export function setMockResponses(m) {
@@ -168,7 +176,9 @@ export function setMockResponses(m) {
 
 export default {
   send: send,
-  receive: receive,
+  receiveMessage: receiveMessage,
+  handle: handle,
   event: event,
-  request: request
+  receive: receive,
+
 }