+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# [3.2.0-next.17](https://github.com/rdkcentral/firebolt-openrpc/compare/v3.2.0-next.16...v3.2.0-next.17) (2025-06-06)
+
+
+### Features
+
+* Remove all optional parameters with null value ([850e26b](https://github.com/rdkcentral/firebolt-openrpc/commit/850e26bb006dd06691cbe88940db58303abee861))
+
+# [3.2.0-next.16](https://github.com/rdkcentral/firebolt-openrpc/compare/v3.2.0-next.15...v3.2.0-next.16) (2025-05-16)
+
+
+### Bug Fixes
+
+* Fix undefined behavior caused by using reinterpret cast on an object with multiple inheritance ([156ac0e](https://github.com/rdkcentral/firebolt-openrpc/commit/156ac0e13c35396083a8b32b839bffdf102dd30b))
+* Json-engine choose correct openrpc ([a8363b0](https://github.com/rdkcentral/firebolt-openrpc/commit/a8363b0ddfec21ef32158ca24cb263ef229257d5))
+
 # [3.2.0-next.15](https://github.com/rdkcentral/firebolt-openrpc/compare/v3.2.0-next.14...v3.2.0-next.15) (2025-02-28)